Interview-Evaluator/requirements.txt
fastapi[standard]
uvicorn
pydantic
python-dotenv
google-cloud-aiplatform